fix!: use type union instead of optional for fields
Changes how nullability is expressed in the generated types.

Previously every field was optional
```ts
type Message = {
  field?: string;
}
```
which is convenient as there is no need to specify every field that
exists in the message when declaring it.

However this have some unfortunate side effects, most notably that
almost every type is assignable to other types.

```ts
type MessageA = {
  foo?: string;
}

type MessageB = {
  bar?: string;
}

const a: MessageA = {}
const b: MessageB = a; // ok because no field needs to be set in
MessageB.
```

This can lead to subtle bugs when passing these types around.

With this change every field is no longer optional, but `undefined` is a
valid value for every field.
```ts
type MessageA = {
  field: string | undefined;
}
```

Fields belonging to a oneof is still typed as optional, because only one
of these fields should ever be set.

```ts
type MessageA = {
  foo: string | undefined;
}

type MessageB = {
  bar: string | undefined;
}

const a: MessageA = { foo: undefined };
const b: MessageB = a; // error
```

---

BREAKING CHANGE: Fields that could previously be omitted when declaring
a variable, now needs to explicitly be set to `undefined`. Code that
reads these fields will not break.