feat: generate behavior comments for fields
`google.api.field_behavior` is used for describing field
behaviors for fields programmatically, instead of in comments. Because
they are not in comments, they wont be reflected in the generated
Typescript code. This change includes an appended behavior comment for
all fields that are annotated with field_behavior.

Proto3 `optional` do not have this comment appended, because the
generated types already sufficiently signals this.