feat: add metadata about service and method to requesthandler
Adding metadata about the service and method which is invoked
to the call of the requestHandler. This allows a request-handler
to resolve proper service and method configurations for http requests
if need be. An example usage is to use the default service configuration
in order to resolve retry- or hedging-policy for a call.

This change should be backwards compatible, since the request handler
signature specifies the meta as an optional argument.